+++ apps/sending-func/src/__mocks__/env-config.mock.ts
@@ -4,28 +4,22 @@ import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { IConfig } from "../utils/config";
 
 export const envConfig: IConfig = {
-  APPINSIGHTS_DISABLE: false,
   APPINSIGHTS_SAMPLING_PERCENTAGE: 5,
   APPLICATIONINSIGHTS_CONNECTION_STRING: "aKey" as NonEmptyString,
 
-  BACKEND_BASE_URL: "aBaseUrl" as NonEmptyString,
-
-  BACKEND_TOKEN: "aToken" as NonEmptyString,
-
   COSMOSDB_NAME: "aName" as NonEmptyString,
   COSMOSDB_URI: "aUri" as NonEmptyString,
-
   INTERNAL_USER_ID: "internalUserId" as NonEmptyString,
+
   IO_COM_PUSH_NOTIFICATIONS_REDIRECT_PERCENTAGE: "0.3" as NonEmptyString,
   MESSAGE_CONTAINER_NAME: "aContainerName" as NonEmptyString,
 
   MESSAGE_CONTENT_STORAGE_CONNECTION_STRING: "aConnString" as NonEmptyString,
   NEW_NOTIFICATION_QUEUE_NAME: "aQueueName" as NonEmptyString,
-
   NEW_NOTIFICATION_QUEUE_STORAGE_CONNECTION_STRING:
     "aConnString" as NonEmptyString,
-  NOTIFICATION_QUEUE_NAME: "aQueueName" as NonEmptyString,
 
+  NOTIFICATION_QUEUE_NAME: "aQueueName" as NonEmptyString,
   NOTIFICATION_QUEUE_STORAGE_CONNECTION_STRING: "aConnString" as NonEmptyString,
 
   RC_CONFIGURATION_CACHE_TTL: 1000 as NonNegativeInteger,
@@ -34,6 +28,9 @@ export const envConfig: IConfig = {
   REMOTE_CONTENT_COSMOSDB_NAME: "aName" as NonEmptyString,
 
   REMOTE_CONTENT_COSMOSDB_URI: "aUri" as NonEmptyString,
+  SESSION_MANAGER_API_KEY: "apiKey" as NonEmptyString,
+
+  SESSION_MANAGER_BASE_URL: "aBaseUrl" as NonEmptyString,
 
   isProduction: false,
 };