+++ infra/resources/_modules/function_app_sending/locals.tf
@@ -34,6 +34,9 @@ locals {
       BACKEND_BASE_URL = "https://io-p-app-appbackendli.azurewebsites.net"
       BACKEND_TOKEN    = var.appbackendli_token
 
+      SESSION_MANAGER_API_KEY  = "@Microsoft.KeyVault(VaultName=${var.key_vault.name};SecretName=session-manager-api-key)",
+      SESSION_MANAGER_BASE_URL = var.session_manager_base_url,
+
       // INTERNAL USE PROPERTIES
       INTERNAL_USER_ID           = var.internal_user_id
       RC_CONFIGURATION_CACHE_TTL = "28800"