+++ apps/cqrs-func/CHANGELOG.md
@@ -1,5 +1,17 @@
 # cqrs-func
 
+## 1.0.2
+
+### Patch Changes
+
+- db2d7e1: Remove azure-files and table from healthcheck
+
+## 1.0.1
+
+### Patch Changes
+
+- 5c3250c: Remove appinsights_disable variable
+
 ## 1.0.0
 
 ### Major Changes